+++ docs/changelog.rst
@@ -281,6 +281,8 @@ Detailed list of changes
 
 - Text sizing protocol: Fix a buffer overflow when a natural width text sizing escape code contains a grapheme cluster longer than four codepoints (:cve:`2026-80431`)
 
+- ssh kitten askpass: Verify owner and permissions of SHM memory used for askpass
+
 - Graphics protocol: Fix a crash when transmitting image data via a file or
   shared memory object (``t=f``, ``t=t`` or ``t=s``) and the client truncates
   it while kitty is reading from it.